+++ src/utils.ts
@@ -3,20 +3,20 @@ import * as fs from 'fs';
 import * as micromatch from 'micromatch';
 import * as path from 'path';
 import * as typescript from 'typescript';
-import * as webpack from 'webpack';
 
 import constants = require('./constants');
 import {
-  DependencyGraph,
   ErrorInfo,
+  FilePathKey,
   LoaderOptions,
+  ResolvedModule,
   ReverseDependencyGraph,
   Severity,
   TSInstance,
   WebpackError,
   WebpackModule,
 } from './interfaces';
-
+import { getInputFileNameFromOutput } from './instances';
 /**
  * The default error formatter.
  */
@@ -111,7 +111,7 @@ export function formatErrors(
         });
 }
 
-export function readFile(
+export function fsReadFile(
   fileName: string,
   encoding: string | undefined = 'utf8'
 ) {
@@ -174,56 +174,65 @@ export function unorderedRemoveItem<T>(array: T[], item: T): boolean {
   return false;
 }
 
+export function populateDependencyGraph(
+  resolvedModules: ResolvedModule[],
+  instance: TSInstance,
+  containingFile: string
+) {
+  resolvedModules = resolvedModules.filter(
+    mod => mod !== null && mod !== undefined
+  );
+  if (resolvedModules.length) {
+    const containingFileKey = instance.filePathKeyMapper(containingFile);
+    instance.dependencyGraph.set(containingFileKey, resolvedModules);
+  }
+}
+
+export function populateReverseDependencyGraph(instance: TSInstance) {
+  const reverseDependencyGraph: ReverseDependencyGraph = new Map();
+  for (const [fileKey, resolvedModules] of instance.dependencyGraph.entries()) {
+    const inputFileName =
+      instance.solutionBuilderHost &&
+      getInputFileNameFromOutput(instance, fileKey);
+    const containingFileKey = inputFileName
+      ? instance.filePathKeyMapper(inputFileName)
+      : fileKey;
+    resolvedModules.forEach(({ resolvedFileName }) => {
+      const key = instance.filePathKeyMapper(
+        instance.solutionBuilderHost
+          ? getInputFileNameFromOutput(instance, resolvedFileName) ||
+              resolvedFileName
+          : resolvedFileName
+      );
+      let map = reverseDependencyGraph.get(key);
+      if (!map) {
+        map = new Map();
+        reverseDependencyGraph.set(key, map);
+      }
+      map.set(containingFileKey, true);
+    });
+  }
+  return reverseDependencyGraph;
+}
+
 /**
  * Recursively collect all possible dependants of passed file
  */
 export function collectAllDependants(
   reverseDependencyGraph: ReverseDependencyGraph,
-  fileName: string,
-  collected: { [file: string]: boolean } = {}
-): string[] {
-  const result = {};
-  result[fileName] = true;
-  collected[fileName] = true;
-  const dependants = reverseDependencyGraph[fileName];
+  fileName: FilePathKey,
+  result: Map<FilePathKey, true> = new Map()
+): Map<FilePathKey, true> {
+  result.set(fileName, true);
+  const dependants = reverseDependencyGraph.get(fileName);
   if (dependants !== undefined) {
-    Object.keys(dependants).forEach(dependantFileName => {
-      if (!collected[dependantFileName]) {
-        collectAllDependants(
-          reverseDependencyGraph,
-          dependantFileName,
-          collected
-        ).forEach(fName => (result[fName] = true));
+    for (const dependantFileName of dependants.keys()) {
+      if (!result.has(dependantFileName)) {
+        collectAllDependants(reverseDependencyGraph, dependantFileName, result);
       }
-    });
+    }
   }
-  return Object.keys(result);
-}
-
-/**
- * Recursively collect all possible dependencies of passed file
- */
-export function collectAllDependencies(
-  dependencyGraph: DependencyGraph,
-  filePath: string,
-  collected: { [file: string]: boolean } = {}
-): string[] {
-  const result = {};
-  result[filePath] = true;
-  collected[filePath] = true;
-  const directDependencies = dependencyGraph[filePath];
-  if (directDependencies !== undefined) {
-    directDependencies.forEach(dependencyModule => {
-      if (!collected[dependencyModule.originalFileName]) {
-        collectAllDependencies(
-          dependencyGraph,
-          dependencyModule.resolvedFileName,
-          collected
-        ).forEach(depFilePath => (result[depFilePath] = true));
-      }
-    });
-  }
-  return Object.keys(result);
+  return result;
 }
 
 export function arrify<T>(val: T | T[]) {
@@ -232,6 +241,22 @@ export function arrify<T>(val: T | T[]) {
   }
 
   return Array.isArray(val) ? val : [val];
+}
+
+export function ensureTrailingDirectorySeparator<T extends string>(dir: T): T {
+  return hasTrailingDirectorySeparator(dir) ? dir : ((dir + '/') as T);
+}
+
+function isAnyDirectorySeparator(charCode: number): boolean {
+  return (
+    charCode === 0x2f || charCode === 0x5c // /
+  ); // \
+}
+
+function hasTrailingDirectorySeparator(dir: string) {
+  return (
+    dir.length > 0 && isAnyDirectorySeparator(dir.charCodeAt(dir.length - 1))
+  );
 }
 
 export function ensureProgram(instance: TSInstance) {
@@ -254,164 +279,29 @@ export function ensureProgram(instance: TSInstance) {
   return instance.program;
 }
 
-export function supportsProjectReferences(instance: TSInstance) {
-  const program = ensureProgram(instance);
-  return program && !!program.getProjectReferences;
-}
-
-export function isUsingProjectReferences(instance: TSInstance) {
-  if (
-    instance.loaderOptions.projectReferences &&
-    supportsProjectReferences(instance)
-  ) {
-    const program = ensureProgram(instance);
-    return Boolean(program && program.getProjectReferences());
-  }
-  return false;
-}
-
-/**
- * Gets the project reference for a file from the cache if it exists,
- * or gets it from TypeScript and caches it otherwise.
- */
-export function getAndCacheProjectReference(
-  filePath: string,
-  instance: TSInstance
-) {
-  // When using solution builder, dont do the project reference caching
-  if (instance.solutionBuilderHost) {
-    return undefined;
-  }
-
-  const file = instance.files.get(filePath);
-  if (file !== undefined && file.projectReference) {
-    return file.projectReference.project;
-  }
-
-  const projectReference = getProjectReferenceForFile(filePath, instance);
-  if (file !== undefined) {
-    file.projectReference = { project: projectReference };
-  }
-
-  return projectReference;
-}
-
-function getResolvedProjectReferences(
-  program: typescript.Program
-): typescript.ResolvedProjectReference[] | undefined {
-  const getProjectReferences =
-    (program as any).getResolvedProjectReferences ||
-    program.getProjectReferences;
-  if (getProjectReferences) {
-    return getProjectReferences();
-  }
-  return;
-}
-
-function getProjectReferenceForFile(filePath: string, instance: TSInstance) {
-  if (isUsingProjectReferences(instance)) {
-    const program = ensureProgram(instance);
-    return (
-      program &&
-      getResolvedProjectReferences(program)!.find(
-        ref =>
-          (ref &&
-            ref.commandLine.fileNames.some(
-              file => path.normalize(file) === filePath
-            )) ||
-          false
-      )
-    );
-  }
-
-  return;
-}
-
-export function validateSourceMapOncePerProject(
-  instance: TSInstance,
-  loader: webpack.loader.LoaderContext,
-  jsFileName: string,
-  project: typescript.ResolvedProjectReference
-) {
-  const { projectsMissingSourceMaps = new Set<string>() } = instance;
-  if (!projectsMissingSourceMaps.has(project.sourceFile.fileName)) {
-    instance.projectsMissingSourceMaps = projectsMissingSourceMaps;
-    projectsMissingSourceMaps.add(project.sourceFile.fileName);
-    const mapFileName = jsFileName + '.map';
-    if (!instance.compiler.sys.fileExists(mapFileName)) {
-      const [relativeJSPath, relativeProjectConfigPath] = [
-        path.relative(loader.rootContext, jsFileName),
-        path.relative(loader.rootContext, project.sourceFile.fileName),
-      ];
-      loader.emitWarning(
-        new Error(
-          'Could not find source map file for referenced project output ' +
-            `${relativeJSPath}. Ensure the 'sourceMap' compiler option ` +
-            `is enabled in ${relativeProjectConfigPath} to ensure Webpack ` +
-            'can map project references to the appropriate source files.'
-        )
-      );
-    }
-  }
-}
-
 export function supportsSolutionBuild(instance: TSInstance) {
   return (
     !!instance.configFilePath &&
     !!instance.loaderOptions.projectReferences &&
-    !!instance.compiler.InvalidatedProjectKind &&
     !!instance.configParseResult.projectReferences &&
     !!instance.configParseResult.projectReferences.length
   );
 }
 
-/**
- * Gets the output JS file path for an input file governed by a composite project.
- * Pulls from the cache if it exists; computes and caches the result otherwise.
- */
-export function getAndCacheOutputJSFileName(
-  inputFileName: string,
-  projectReference: typescript.ResolvedProjectReference,
-  instance: TSInstance
-) {
-  const file = instance.files.get(inputFileName);
-  if (file && file.projectReference && file.projectReference.outputFileName) {
-    return file.projectReference.outputFileName;
-  }
-
-  const outputFileName = getOutputJavaScriptFileName(
-    inputFileName,
-    projectReference
+export function isReferencedFile(instance: TSInstance, filePath: string) {
+  return (
+    !!instance.solutionBuilderHost &&
+    !!instance.solutionBuilderHost.watchedFiles.get(
+      instance.filePathKeyMapper(filePath)
+    )
   );
-
-  if (file !== undefined) {
-    file.projectReference = file.projectReference || {
-      project: projectReference,
-    };
-    file.projectReference.outputFileName = outputFileName;
-  }
-
-  return outputFileName;
 }
 
-// Adapted from https://github.com/Microsoft/TypeScript/blob/45101491c0b077c509b25830ef0ee5f85b293754/src/compiler/tsbuild.ts#L305
-function getOutputJavaScriptFileName(
-  inputFileName: string,
-  projectReference: typescript.ResolvedProjectReference
+export function useCaseSensitiveFileNames(
+  compiler: typeof typescript,
+  loaderOptions: LoaderOptions
 ) {
-  const { options } = projectReference.commandLine;
-  const projectDirectory =
-    options.rootDir || path.dirname(projectReference.sourceFile.fileName);
-  const relativePath = path.relative(projectDirectory, inputFileName);
-  const outputPath = path.resolve(
-    options.outDir || projectDirectory,
-    relativePath
-  );
-  const newExtension = constants.jsonRegex.test(inputFileName)
-    ? '.json'
-    : constants.tsxRegex.test(inputFileName) &&
-      options.jsx === typescript.JsxEmit.Preserve
-    ? '.jsx'
-    : '.js';
-  return outputPath.replace(constants.extensionRegex, newExtension);
+  return loaderOptions.useCaseSensitiveFileNames !== undefined
+    ? loaderOptions.useCaseSensitiveFileNames
+    : compiler.sys.useCaseSensitiveFileNames;
 }