+++ test/comparison-tests/projectReferencesMultiple/expectedOutput-3.9/patch0/lib/fileWithError.js
@@ -1,0 +1,7 @@
+"use strict";
+exports.__esModule = true;
+exports.foo = void 0;
+function foo() {
+    return "hello world";
+}
+exports.foo = foo;