+++ test/comparison-tests/projectReferencesSymLinks/lib/src/index.ts
@@ -1,0 +1,2 @@
+import { getMeaningOfLife2 } from "common";
+export const getMeaningOfLife = () => getMeaningOfLife2();