+++ test/comparison-tests/projectReferencesWatchRefWithTwoFilesAlreadyBuilt/expectedOutput-3.9/patch1/lib/helper.js
@@ -1,0 +1,10 @@
+"use strict";
+exports.__esModule = true;
+exports.helper = void 0;
+exports.helper = {
+    one: 1,
+    two: 2,
+    three: 3,
+    four: 4
+};
+//# sourceMappingURL=helper.js.map