+++ test/comparison-tests/projectReferencesWatchRefWithTwoFilesAlreadyBuilt/expectedOutput-transpile-3.9/patch1/lib/helper.d.ts
@@ -1,0 +1,6 @@
+export declare const helper: {
+    one: number;
+    two: number;
+    three: number;
+    four: number;
+};